requirements.txt
opencv-python-headless>=4.5.0
face_recognition>=1.3.0
numpy>=1.19.0
Pillow>=8.0.0
pytesseract>=0.3.8
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
websockets>=11.0
